Move callout styling into CampaignCalloutListSpec
